+++ BracketMap.DAL/Entities/Player.cs
@@ -1,0 +1,8 @@
+﻿namespace BracketMap.DAL.Entities
+{
+    public class Player
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}
