+++ BracketMap.DAL/Entities/Tournament.cs
@@ -1,0 +1,11 @@
+﻿namespace BracketMap.DAL.Entities
+{
+    public class Tournament
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public int PlayerCount { get; set; }
+        public int TeamCount { get; set; }
+        public string Status { get; set; }
+    }
+}
